Fixes for issue gh-15:
Show/Hide the "Add Additional Location" button according to whether or
not the user has selected coordinates for the currently active
location.

Prevent adding an additional location when the user hasn't checked any
ecosystems for the active location and warn the user.

Add some minimal instructions for using the calculator and revise and
clarify existing text.