Remove netcdf files from repo
I'm putting all of the netcdf files in a single, central location for
now. Will add a script to download the package of the files &
automatically extract them to the right folder later

Doing this to try and lean up the `git clone` command, but also to
centralize the files.